Replace protobuf messages with hand-written domain types
The generated protobuf structs have been serving as this service's domain
types: registry.Skill.Metadata was a *skillsv1.SkillMetadata, and
proposals.ProposeChange took a request message directly. That was fine
while gRPC was the only transport. It stops being fine under MCP, where
tool schemas are inferred from Go types and protoc-gen-go's unconditional
`json:",omitempty"` on every field would make every argument optional.

Introduce internal/skill for the metadata types, give internal/proposals
its own Proposal/Commit/Endorsement/FileEdit/Cluster/Submission, and add a
string wire form for evidence.Verdict. gRPC still works: internal/protomap
maps domain types back to the generated messages, and dies with them.

Three things worth calling out beyond the mechanical conversion:

proto.Clone is gone, replaced by skill.Metadata.Clone. This is the one
change here that the compiler cannot check. The registry's copy of a skill
is shared across every concurrent request on a replica, so a clone that
aliases the metadata map or the context-file slice corrupts the index for
every later request - intermittently, under load, looking like a caching
bug. ContextFiles is []ContextFile by value rather than []*ContextFile so
the pointer-copying version of that mistake is unwritable, and two tests
pin the rest: one fails if a field is added to Metadata without extending
Clone, and one hammers the shared entry from 100 goroutines and asserts
the index is untouched afterward.

The evidence store's verdict integers are frozen and now say so. They are
written to SQLite and form part of the signal_rollup primary key, so the
wire form is a string and the mapping is explicit. Verdicts replaces the
contiguous `for v := X; v <= Y; v++` walk that signal serialization
depended on - the ordering guarantee was real but untested.

ListProposals and ListClusters no longer compute a diff per branch. Both
went through GetProposal, which does a tree-to-tree comparison; clustering
never read the result, and a listing's job is to help pick which proposal
to look at, not to render all of them.

+++ internal/registryconfig/registryconfig.go
@@ -84,7 +84,7 @@ type Config struct {
 	// identities, one misbehaving caller can manufacture a threshold's
 	// worth of agreement by itself. Enable it once callers are
 	// authenticated, not before.
-	AutoSubmitEndorsements int32
+	AutoSubmitEndorsements int
 
 	// FetchInterval is how often the base branch is re-fetched from
 	// origin in the background.
@@ -208,7 +208,7 @@ func Load() (Config, error) {
 		GitHubOwner:             getenv("GITHUB_OWNER", ""),
 		GitHubRepo:              getenv("GITHUB_REPO", ""),
 		GitHubAPIBaseURL:        getenv("GITHUB_API_BASE_URL", "https://api.github.com"),
-		AutoSubmitEndorsements:  int32(autoSubmitEndorsements),
+		AutoSubmitEndorsements:  autoSubmitEndorsements,
 		FetchInterval:           fetchInterval,
 		EvidenceEnabled:         evidenceEnabled,
 		EvidenceDBPath:          getenv("EVIDENCE_DB_PATH", "/var/lib/skillsd-evidence/evidence.db"),
